+++ app/Widgets/CommunityData/CommunityData.php
@@ -5,6 +5,7 @@ namespace App\Widgets\CommunityData;
 use Moxl\Xec\Action\Pubsub\GetItem;
 
 use Movim\Widget\Base;
+use Moxl\Xec\Payload\Packet;
 
 class CommunityData extends Base
 {
@@ -13,25 +14,32 @@ class CommunityData extends Base
         $this->addcss('communitydata.css');
         $this->addjs('communitydata.js');
         $this->registerEvent('disco_request_handle', 'onDiscoRequest', 'community');
-        $this->registerEvent('pubsub_getitem_avatar', 'onDiscoRequest', 'community');
+        $this->registerEvent('pubsub_getitem_avatar', 'onAvatar', 'community');
     }
 
-    public function onDiscoRequest($packet)
+    public function onDiscoRequest(Packet $packet)
+    {
+        $info = $packet->content;
+
+        if (!$info->isMicroblogCommentsNode()) {
+            $this->rpc('MovimTpl.fill', '#community_data', $this->prepareData($info->server, $info->node));
+        }
+    }
+
+    public function onAvatar(Packet $packet)
     {
         list($origin, $node) = array_values($packet->content);
 
-        if ((substr($node, 0, 30) != 'urn:xmpp:microblog:0:comments/')) {
-            $this->rpc('MovimTpl.fill', '#community_data', $this->prepareData($origin, $node));
-        }
+        $this->rpc('MovimTpl.fill', '#community_data', $this->prepareData($origin, $node));
     }
 
     public function ajaxGetAvatar($origin, $node)
     {
         $g = new GetItem;
         $g->setTo($origin)
-          ->setNode($node)
-          ->setId('urn:xmpp:avatar:metadata')
-          ->request();
+            ->setNode($node)
+            ->setId('urn:xmpp:avatar:metadata')
+            ->request();
     }
 
     public function prepareCard($info)
@@ -41,12 +49,13 @@ class CommunityData extends Base
         $view->assign('num', 0);
 
         if ($info) {
-            $view->assign('num',
+            $view->assign(
+                'num',
                 ($info->items > 0)
                     ? $info->items
                     : \App\Post::where('server', $info->server)
-                           ->where('node', $info->node)
-                           ->count()
+                    ->where('node', $info->node)
+                    ->count()
             );
         } else {
             return '';
@@ -59,23 +68,25 @@ class CommunityData extends Base
     {
         $view = $this->tpl();
         $info = \App\Info::where('server', $origin)
-                         ->where('node', $node)
-                         ->first();
+            ->where('node', $node)
+            ->first();
 
         $view->assign('info', $info);
         $view->assign('num', 0);
 
         if ($info) {
-            $view->assign('num',
+            $view->assign(
+                'num',
                 ($info->items > 0)
                     ? $info->items
                     : \App\Post::where('server', $origin)
-                           ->where('node', $node)
-                           ->count()
+                    ->where('node', $node)
+                    ->count()
             );
 
             $title = !empty($info->name) ? $info->name : $node;
-            $this->rpc('Notif.setTitle',
+            $this->rpc(
+                'Notif.setTitle',
                 $this->__('page.communities') . ' • ' . $title
             );
         }