+++ src/Moxl/Xec/Action/Bookmark2/Get.php
@@ -19,8 +19,7 @@ class Get extends Action
 
     public function handle(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null)
     {
-        me()
-            ->session
+        me()->session
             ->conferences()
             ->where('bookmarkversion', (int)$this->_version)
             ->delete();
@@ -36,8 +35,7 @@ class Get extends Action
         }
 
         // We remove the conferences that might be saved under another bookmark version
-        me()
-            ->session
+        me()->session
             ->conferences()
             ->whereIn('conference', $conferenceIds)
             ->delete();